+++ Src/Library/Constants.cs
@@ -7,7 +7,7 @@ internal static class Constants
     internal const string PermissionsClaimType = "permissions";
 
     //this is used as a dictionary key. int32.gethashcode() just returns the value
-    internal const int ResponseSent = 0;
+    internal const byte ResponseSent = 0;
 
     //must match Microsoft.AspNetCore.Http.ProducesResponseTypeMetadata sealed class
     internal const string ProducesMetadata = "ProducesResponseTypeMetadata";